Edit response message

+++ TryJWT/src/main/java/com/Quan/TryJWT/controller/BrandController.java
@@ -7,6 +7,7 @@ import java.util.Optional;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -19,6 +20,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.Quan.TryJWT.Exception.AppUtils;
 import com.Quan.TryJWT.Exception.NotFoundException;
 import com.Quan.TryJWT.dto.BrandOutput;
 import com.Quan.TryJWT.model.Brand;
@@ -78,10 +80,11 @@ public class BrandController {
 	@GetMapping(value = { "/{id}" })
 	public ResponseEntity<?> getBrandById(@PathVariable("id") long id) {
 		Brand brand = null;
-		try {
-			brand = brandService.findById(id);
-		} catch (NotFoundException e) {
-			return ResponseEntity.badRequest().body("Brand is unavaiable");
+
+		brand = brandService.findById(id);
+		
+		if(brand == null) {
+			return AppUtils.returnJS(HttpStatus.BAD_REQUEST, "Brand is unavaiable", null);
 		}
 		return ResponseEntity.ok(brand);
 	}
@@ -89,52 +92,49 @@ public class BrandController {
 	@PostMapping
 	public ResponseEntity<?> postBrand(@Valid @RequestBody Brand brand, BindingResult bindingResult) {
 		if (brandService.existsByName(brand.getName())) {
-			return ResponseEntity
-					.badRequest()
-					.body("Error: Brand name is already taken!");
+			return AppUtils.returnJS(HttpStatus.BAD_REQUEST, "Brand name is already taken!", null);
 		}
 		if (bindingResult.hasErrors())
-			return ResponseEntity
-					.badRequest()
-					.body("Error: " + bindingResult.getAllErrors().get(0).getDefaultMessage());
-
+			return AppUtils.returnJS(HttpStatus.BAD_REQUEST, bindingResult.getAllErrors().get(0).getDefaultMessage(), null);
+		
 		brandService.addBrand(brand);
-		return ResponseEntity.ok("Add brand successfully!");
+		
+		return AppUtils.returnJS(HttpStatus.OK, "Add brand successfully!", brand);
 	}
 	
 	@DeleteMapping(value = "/{id}")
 	public ResponseEntity<?> deleteBrand(@PathVariable("id") long id) {
 		Brand brand = null;
-		try {
-			brand = brandService.findById(id);
-			if(brand.getProducts().size() > 0)
-				return ResponseEntity.badRequest().body("Brand is in use");
-		} catch (NotFoundException e) {
-			return ResponseEntity.badRequest().body("Brand is unavaiable");
+
+		brand = brandService.findById(id);
+		
+		if(brand == null) {
+			return AppUtils.returnJS(HttpStatus.BAD_REQUEST, "Brand is unavaiable", null);
 		}
+		
+		if(brand.getProducts().size() > 0)
+			return AppUtils.returnJS(HttpStatus.BAD_REQUEST, "Brand is in use", null);
+
 		brandService.deleteBrand(brand);
-		return ResponseEntity.ok("Remove brand successfully!");
+		return AppUtils.returnJS(HttpStatus.OK, "Remove brand successfully!", null);
 	}
 	
 	@PutMapping(value = "/{id}")
 	public ResponseEntity<?> putBrand(@PathVariable("id") long id, @Valid @RequestBody Brand newbrand, BindingResult bindingResult) {
 		Brand oldBrand = null;
-		try {
-			oldBrand = brandService.findById(id);
-			if (brandService.existsByName(newbrand.getName()) && !newbrand.getName().equals(oldBrand.getName())) {
-				return ResponseEntity
-						.badRequest()
-						.body("Error: Brand name is already taken!");
-			}
-			if (bindingResult.hasErrors())
-				return ResponseEntity
-						.badRequest()
-						.body("Error: " + bindingResult.getAllErrors().get(0).getDefaultMessage());
-		} catch (NotFoundException e) {
-			return ResponseEntity.badRequest().body("Brand is unavaiable");
-		}
+		oldBrand = brandService.findById(id);
 		
+		if(oldBrand == null)
+			return AppUtils.returnJS(HttpStatus.BAD_REQUEST, "Brand is unavaiable", null);
+		
+		if (brandService.existsByName(newbrand.getName()) && !newbrand.getName().equals(oldBrand.getName())) {
+			return AppUtils.returnJS(HttpStatus.BAD_REQUEST, "Brand name is already taken!", null);
+		}
+		if (bindingResult.hasErrors())
+			return AppUtils.returnJS(HttpStatus.BAD_REQUEST, bindingResult.getAllErrors().get(0).getDefaultMessage(), null);
+		
+		newbrand.setBrandId(id);
 		brandService.updateBrand(newbrand);
-		return ResponseEntity.ok("Update brand successfully!");
+		return AppUtils.returnJS(HttpStatus.OK, "Update brand successfully!", newbrand);
 	}
 }
